arm64: Use the new MULTI_IRQ_HANDLER
It appears arm64 copied arm's MULTI_IRQ_HANDLER code, but made it
unconditional.  I wanted to make this generic so it could be used by the
RISC-V port.  This patch converts the arm64 code to use the new generic
code, which simply consists of deleting the arm64 code and setting
MULTI_IRQ_HANDLER instead.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>